requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.3
exceptiongroup==1.1.2
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
gradio_client==0.16.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
# huggingface-hub==0.16.4
huggingface-hub
idna==3.4
multidict==6.0.4
openai==0.27.8
packaging==23.1
# PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
websockets==11.0.3
yarl==1.9.2
python-dotenv==1.0.0
